+++ tests/phpunit/Admin_Test.php
@@ -1282,7 +1282,6 @@ class Admin_Test extends WP_StreamTestCase {
 			public $locations;
 			public $admin;
 			public $connectors;
-			public $network;
 			public $scheduler;
 			public function __construct( $real ) {
 				$this->settings   = $real->settings;
@@ -1290,7 +1289,6 @@ class Admin_Test extends WP_StreamTestCase {
 				$this->locations  = $real->locations;
 				$this->admin      = $real->admin;
 				$this->connectors = $real->connectors;
-				$this->network    = $real->network;
 				$this->scheduler  = $real->scheduler;
 			}
 			public function is_multisite_not_network_activated() {